Remove throttle and rudder feedback from Arduino
I originally designed the Arduino interface to work both ways: new data was
sent to the Arduino, and it would return their actual current state. This was a
mistake, as the navigation loop uses the cross-track distance as the feedback
variable, and the rudder is independently controlled with PID on the Arduino.

This hugely simplifies the integration of pilot and navigation modules, as
communication is now exclusively one-way.